# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ManagementSystem API",
        "description": "API documentation for ManagementSystem",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/ManagementSystem/public",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/tasks/project/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get all tasks for a project",
                "operationId": "70366d2b9f564e3d5e009120e63dd504",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tasks retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tasks Project retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TaskResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Project not found"
                    },
                    "429": {
                        "description": "Too Many Requests - Throttled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tasks/show/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get a single task",
                "operationId": "f932c5f0499bffd52816d60f3ebc193e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Task retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TaskResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Task not found"
                    },
                    "429": {
                        "description": "Too Many Requests - Throttled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tasks/create/{ProjectId}/{UserId}": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create a new task for a user in a project",
                "operationId": "a3f39905d41852bdd273ac47e23db921",
                "parameters": [
                    {
                        "name": "ProjectId",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "UserId",
                        "in": "path",
                        "description": "User ID to assign the task to",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Implement authentication",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Implement JWT authentication for the API endpoints"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Task created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized - Admin or Owner role required"
                    },
                    "404": {
                        "description": "Project or User not found / User not assigned to project"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "429": {
                        "description": "Too Many Requests - Throttled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tasks/change-status/{TaskId}/{StatusId}": {
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Change task status",
                "operationId": "78e2fcc36021a4baac22e23ab4631837",
                "parameters": [
                    {
                        "name": "TaskId",
                        "in": "path",
                        "description": "Task ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "StatusId",
                        "in": "path",
                        "description": "New Status ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Task status updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Task or Status not found"
                    },
                    "429": {
                        "description": "Too Many Requests - Throttled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tasks/{id}": {
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete a task",
                "operationId": "25647b38472bf8c884b793a598b2c990",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Task deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized - Admin or Owner role required"
                    },
                    "404": {
                        "description": "Task not found"
                    },
                    "429": {
                        "description": "Too Many Requests - Throttled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProjectResource": {
                "title": "ProjectResource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "New Project"
                    },
                    "description": {
                        "type": "string",
                        "example": "new project description",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-20T08:30:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-20T09:15:00.000000Z"
                    },
                    "users": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserResource"
                        },
                        "nullable": true
                    },
                    "tasks": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TaskResource"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "RoleResource": {
                "title": "RoleResource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "admin"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-20T08:30:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-20T09:15:00.000000Z"
                    },
                    "users": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserResource"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "StatusResource": {
                "title": "StatusResource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "pending"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-20T08:30:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-20T09:15:00.000000Z"
                    },
                    "tasks": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TaskResource"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "TaskResource": {
                "title": "TaskResource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "title task"
                    },
                    "description": {
                        "type": "string",
                        "example": "task description",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-20T08:30:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-20T09:15:00.000000Z"
                    },
                    "user": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/UserResource"
                            }
                        ],
                        "nullable": true,
                        "type": "object"
                    },
                    "project": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/ProjectResource"
                            }
                        ],
                        "nullable": true,
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "title": "UserResource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Ahmed Ali"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "ahmed@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+966501234567",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-20T08:30:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-20T09:15:00.000000Z"
                    },
                    "active": {
                        "type": "boolean",
                        "example": true
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RoleResource"
                        },
                        "nullable": true
                    },
                    "projects": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProjectResource"
                        },
                        "nullable": true
                    },
                    "statuses": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StatusResource"
                        },
                        "nullable": true
                    },
                    "tasks": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TaskResource"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}